Restructure meeting model and storage
