+++ src/main/java/seedu/address/MainApp.java
@@ -19,23 +19,23 @@ import seedu.address.logic.LogicMode;
 import seedu.address.logic.LogicModeManager;
 import seedu.address.logic.LogicPerson;
 import seedu.address.logic.LogicPersonManager;
-import seedu.address.model.ModelMeeting;
-import seedu.address.model.ModelMeetingManager;
-import seedu.address.model.ModelPerson;
-import seedu.address.model.ModelPersonManager;
+import seedu.address.model.meeting.ModelMeeting;
+import seedu.address.model.meeting.ModelMeetingManager;
+import seedu.address.model.person.ModelPerson;
+import seedu.address.model.person.ModelPersonManager;
 import seedu.address.model.meeting.MeetingBook;
 import seedu.address.model.person.AddressBook;
 import seedu.address.model.person.ReadOnlyAddressBook;
-import seedu.address.model.person.ReadOnlyUserPrefs;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.ReadOnlyUserPrefs;
+import seedu.address.model.UserPrefs;
 import seedu.address.model.util.SampleDataUtil;
-import seedu.address.storage.StorageMeeting;
-import seedu.address.storage.StoragePerson;
-import seedu.address.storage.StoragePersonManager;
+import seedu.address.storage.meeting.StorageMeeting;
+import seedu.address.storage.person.StoragePerson;
+import seedu.address.storage.person.StoragePersonManager;
 import seedu.address.storage.person.AddressBookStorage;
 import seedu.address.storage.person.JsonAddressBookStorage;
-import seedu.address.storage.person.JsonUserPrefsStorage;
-import seedu.address.storage.person.UserPrefsStorage;
+import seedu.address.storage.JsonUserPrefsStorage;
+import seedu.address.storage.UserPrefsStorage;
 import seedu.address.ui.Ui;
 import seedu.address.ui.UiManager;
