+++ src/main/java/seedu/address/logic/commands/meeting/AddCommand.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.ModelMeeting;
+import seedu.address.model.meeting.ModelMeeting;
 import seedu.address.model.meeting.Meeting;
 
 public class AddCommand extends Command {
