+++ src/main/java/seedu/address/logic/commands/person/AddCommand.java
@@ -9,7 +9,7 @@ import static seedu.address.logic.parser.person.CliSyntax.PREFIX_TAG;
 
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.ModelPerson;
+import seedu.address.model.person.ModelPerson;
 import seedu.address.model.person.person.Person;
 
 /**
