+++ src/main/java/seedu/address/logic/commands/person/Command.java
@@ -2,7 +2,7 @@ package seedu.address.logic.commands.person;
 
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.ModelPerson;
+import seedu.address.model.person.ModelPerson;
 
 /**
  * Represents a command with hidden internal logic and the ability to be executed.
