+++ src/main/java/seedu/address/logic/commands/person/DeleteCommand.java
@@ -8,7 +8,7 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.ModelPerson;
+import seedu.address.model.person.ModelPerson;
 import seedu.address.model.person.person.Person;
 
 /**
