+++ src/main/java/seedu/address/storage/JsonUserPrefsStorage.java
@@ -1,4 +1,4 @@
-package seedu.address.storage.person;
+package seedu.address.storage;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -6,8 +6,8 @@ import java.util.Optional;
 
 import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.commons.util.JsonUtil;
-import seedu.address.model.person.ReadOnlyUserPrefs;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.ReadOnlyUserPrefs;
+import seedu.address.model.UserPrefs;
 
 /**
  * A class to access UserPrefs stored in the hard disk as a json file
