+++ src/main/java/seedu/address/storage/UserPrefsStorage.java
@@ -1,12 +1,12 @@
-package seedu.address.storage.person;
+package seedu.address.storage;
 
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.Optional;
 
 import seedu.address.commons.exceptions.DataConversionException;
-import seedu.address.model.person.ReadOnlyUserPrefs;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.ReadOnlyUserPrefs;
+import seedu.address.model.UserPrefs;
 
 /**
  * Represents a storage for {@link UserPrefs}.
