+++ src/main/java/seedu/address/storage/person/StoragePerson.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.storage.person;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -6,16 +6,15 @@ import java.util.Optional;
 
 import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.model.person.ReadOnlyAddressBook;
-import seedu.address.model.person.ReadOnlyUserPrefs;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.ReadOnlyUserPrefs;
+import seedu.address.model.UserPrefs;
+import seedu.address.storage.UserPrefsStorage;
 import seedu.address.storage.person.AddressBookStorage;
-import seedu.address.storage.person.UserPrefsStorage;
 
 /**
  * API of the Storage component
  */
-// TODO: build and change to MeetingBookStorage
-public interface StorageMeeting extends AddressBookStorage, UserPrefsStorage {
+public interface StoragePerson extends AddressBookStorage, UserPrefsStorage {
 
     @Override
     Optional<UserPrefs> readUserPrefs() throws DataConversionException, IOException;
