+++ src/test/java/seedu/address/logic/LogicPersonManagerTest.java
@@ -22,14 +22,14 @@ import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.commands.person.AddCommand;
 import seedu.address.logic.commands.person.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.ModelPerson;
-import seedu.address.model.ModelPersonManager;
+import seedu.address.model.person.ModelPerson;
+import seedu.address.model.person.ModelPersonManager;
 import seedu.address.model.person.ReadOnlyAddressBook;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.UserPrefs;
 import seedu.address.model.person.person.Person;
-import seedu.address.storage.StoragePersonManager;
+import seedu.address.storage.person.StoragePersonManager;
 import seedu.address.storage.person.JsonAddressBookStorage;
-import seedu.address.storage.person.JsonUserPrefsStorage;
+import seedu.address.storage.JsonUserPrefsStorage;
 import seedu.address.testutil.PersonBuilder;
 
 public class LogicPersonManagerTest {
