+++ src/test/java/seedu/address/logic/commands/person/ClearCommandTest.java
@@ -5,10 +5,10 @@ import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.model.ModelPerson;
-import seedu.address.model.ModelPersonManager;
+import seedu.address.model.person.ModelPerson;
+import seedu.address.model.person.ModelPersonManager;
 import seedu.address.model.person.AddressBook;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.UserPrefs;
 
 public class ClearCommandTest {
