+++ src/test/java/seedu/address/logic/commands/person/CommandTestUtil.java
@@ -16,7 +16,7 @@ import java.util.List;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.ModelPerson;
+import seedu.address.model.person.ModelPerson;
 import seedu.address.model.person.AddressBook;
 import seedu.address.model.person.person.NameContainsKeywordsPredicate;
 import seedu.address.model.person.person.Person;
