+++ src/test/java/seedu/address/logic/commands/person/DeleteCommandTest.java
@@ -13,9 +13,9 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
-import seedu.address.model.ModelPerson;
-import seedu.address.model.ModelPersonManager;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.person.ModelPerson;
+import seedu.address.model.person.ModelPersonManager;
+import seedu.address.model.UserPrefs;
 import seedu.address.model.person.person.Person;
 
 /**
