+++ src/test/java/seedu/address/model/ModelPersonManagerTest.java
@@ -3,7 +3,7 @@ package seedu.address.model;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.model.ModelPerson.PREDICATE_SHOW_ALL_PERSONS;
+import static seedu.address.model.person.ModelPerson.PREDICATE_SHOW_ALL_PERSONS;
 import static seedu.address.testutil.Assert.assertThrows;
 import static seedu.address.testutil.TypicalPersons.ALICE;
 import static seedu.address.testutil.TypicalPersons.BENSON;
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.model.person.AddressBook;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.person.ModelPersonManager;
 import seedu.address.model.person.person.NameContainsKeywordsPredicate;
 import seedu.address.testutil.AddressBookBuilder;
