+++ src/test/java/seedu/address/model/UserPrefsTest.java
@@ -4,8 +4,6 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.model.person.UserPrefs;
-
 public class UserPrefsTest {
 
     @Test
